1.轨迹信息展示调整 （1）跟轨迹相关的各项查询列表添加GPS是否有效字段 （2）监控主界面添加最后有效轨迹相关字段，轨迹按最后有效轨迹展示 （3）APP获取加油站配送信息接口，返回的轨迹信息添加最后有效轨迹相关字段 （4）监控界面地图上GPS无效时，车辆图标灰色显示 （5）实时监控udp上报轨迹做时间判断，保证展示的轨迹是最新的 （6）重点监控时轨迹列表根据轨迹采集时间间隔播放 2.轨迹回放功能优化 （1）查询的轨迹信息列表可最小化到右下角 （2）轨迹回访时，地图上不展示无效轨迹点 （3）添加是否显示可疑停车地点开关 3.用户中心数据库部分表字段调整 （1）tbl_transport、tbl_remote_alarm_eliminate、tbl_remote_change_gasstation、tbl_remote_control、tbl_unlock_reset添加device_id字段，相关操作添加车台设备ID录入（道闸接口tbl_remote_control表，也添加车台设备ID录入） （2）tbl_remote_control增加lock_id_list字段，远程操作进入应急或是强制变更车辆状态为应急时，添加锁记录ID录入 4.物流配送接口优化 （1）WSDL中隐藏服务接口中非服务方法 （2）修复webservice发布后无法同时使用主机IP和localhost、127.0.0.1同时访问的问题 （3）修复远程换站未成功，物流接口又对同一配送单换站时获取的配送记录ID是无效ID问题 5.信息查询功能优化 （1）查询页面添加加载动画 （2）优化前端分页信息栏处理逻辑，加快页面加载速度 （3）跟锁有关的记录添加车台上报时间默认按车台上报时间查询 6.bug修复 （1）实时监控页面超时未关闭（前端未见超时日志），页面添加计时器，超时主动关闭 （2）未消除报警列表查不到车台报警，修正SQL语句多表关联逻辑 7.新功能添加 （1）sqlite同步失败邮件通知 （2）车台配置页面添加车台事件记录查询 （3）用户中心添加车台事件和开锁重置查询 8.权限更新

+++ src/main/java/com/tipray/bean/alarm/AlarmDevice.java
@@ -13,7 +13,7 @@ public class AlarmDevice implements Serializable {
     /**
      * 车辆ID
      */
-    private int vehicleId;
+    private long vehicleId;
     /**
      * 设备类型（1 车台，2 锁）
      */
@@ -31,11 +31,11 @@ public class AlarmDevice implements Serializable {
      */
     private int alarmType;
 
-    public int getVehicleId() {
+    public long getVehicleId() {
         return vehicleId;
     }
 
-    public void setVehicleId(int vehicleId) {
+    public void setVehicleId(long vehicleId) {
         this.vehicleId = vehicleId;
     }
